Fix incorrect inverse computation modulo (x-1)
In @897f0371 an error was introduced to the key generation routine that
caused a multiple of Phi = (x^N-1)/(x-1) to be added to all inverse
calculations. Let 'b' be the output of the inversion-mod-q routine on
input f. The effect of the bug is that

b mod (x-1) = k (mod q)

where k is related to the number of ones in f^(-1) mod 2. The correct
value for b mod (x-1) is 1.

The exact value of k is given by T(T(T(T(c)))) where
T(x) = -x^2 + (2 - N)*x (mod q)
and c is the number of ones in f^(-1) mod 2. The four-fold
iteration of T is 2-to-1 on Z/qZ, and 1-to-1 on the set of even
elements of Z/qZ. Since c is odd it can be recovered uniquely from k.

Since h(1) = g(1)*b(1) = 1*k, the ones count of f^(-1) mod 2 is
leaked through the public key. This gives the adversary a small
advantage in some combinatorial attacks.

Note that the correct inverse was always computed modulo Phi, and
all existing parameter sets use blinding polynomials, r, such that
r(1) = 0 during encryption. Consequently there are no adverse effects
other than the leakage of the ones count.